Movido rewindTransactions para dentro do index da bitcoin
