+++ external_modules/bitcoin/index.ts
@@ -1,7 +1,8 @@
 import express from 'express'
 import bodyParser from 'body-parser'
 import Common from '../common'
-import Transaction from '../common/db/models/newTransactions'
+import Meta from '../common/db/models/meta'
+import Transaction, { Receive } from '../common/db/models/newTransactions'
 import * as rpc from './methods/rpc'
 import * as methods from './methods'
 import type { WithdrawRequest, WithdrawResponse } from '../common'
@@ -18,8 +19,6 @@ export class Bitcoin extends Common {
 	 * bloqueando novas execuções do rewind
 	 */
 	protected rewinding: boolean
-
-	rewindTransactions = methods.rewindTransactions
 
 	getNewAccount = rpc.getNewAddress
 
@@ -75,6 +74,54 @@ export class Bitcoin extends Common {
 		} catch (err) {
 			console.error('Error fetching unconfirmed transactions', err)
 		}
+	}
+
+	async rewindTransactions(newestBlockhash: string) {
+		this.rewinding = true
+
+		/** Se não encontrar no banco, default para primeiro blockhash válido */
+		const lastKnowHash = await Meta.findOne({ info: 'lastKnowHash' })
+			.map(doc => doc?.details || '000000000933ea01ad0ee984209779baaec3ced90fa3f408719526f8d77f4943')
+
+		const { transactions } = await rpc.listSinceBlock(lastKnowHash)
+
+		const received = await Receive.find({
+			txid: {
+				$in: transactions.map(tx => tx.txid)
+			}
+		})
+
+		/**
+		 * Filtra transações cuja combinação txid & account exista no DB; A account
+		 * é checada não descartar transações em batch recebidas em que uma tá no
+		 * banco e a outra misteriosamente não está
+		 */
+		const txs = transactions.filter(tx =>
+			received.findIndex(s => s.txid == tx.txid && s.account == tx.address) == -1
+		)
+
+		for (const tx of txs) {
+			await this.newTransaction({
+				txid:          tx.txid,
+				status:        tx.confirmations < 3 ? 'pending' : 'confirmed',
+				confirmations: tx.confirmations,
+				account:       tx.address,
+				amount:        tx.amount,
+				timestamp:     tx.time * 1000 // O timestamp do bitcoin é em segundos
+			})
+		}
+
+		await Meta.updateOne({
+			info: 'lastKnowHash'
+		}, {
+			details: newestBlockhash
+		}, {
+			upsert: true
+		})
+
+		const { height } = await rpc.getBlockInfo(newestBlockhash)
+		this.blockHeight = height
+		this.rewinding = false
 	}
 
 	async initBlockchainListener() {
