+++ external_modules/bitcoin/methods/index.ts
@@ -1,4 +1,3 @@
 export * from './processTransaction'
-export * from './rewindTransactions'
 import * as rpc from './rpc'
 export { rpc } // Export tudo de './rpc' dentro de um objeto 'rpc'
